requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
greenlet==3.0.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
testcontainers==3.7.1
anyio==3.7.1
trio==0.23.1
aiosqlite==0.19.0

# Linting and formatting
black==23.11.0
isort==5.12.0
flake8==6.1.0
pre-commit==3.6.0

# Development
python-dotenv==1.0.0 